requirements.txt
# === Main Application ===
fastapi
gunicorn
uvicorn
pydantic[email]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
# === Database ===
SQLAlchemy
psycopg2-binary
alembic

# === Cloud / APIs ===
boto3
google-auth
google-auth-oauthlib
groq

# === CRITICAL VERSION CONSTRAINTS ===
# These specific versions are required to make spleeter, tensorflow, and numpy compatible
tensorflow==2.9.3
numpy==1.23.5
urllib3<2.0

# === Audio / ML Libraries ===
librosa
spleeter
pydub
soundfile

# === Downloaders ===
spotdl
yt-dlp
